+++ src/components/MainSection/MainPage.vue
@@ -1,10 +1,12 @@
 <script setup lang="ts">
 import { PageSelection, type IMasterwork, type IMod, type IPerkOption, type ISelectedGear, type PerkColumnNumber } from '@/data/interfaces';
 import { computed } from '@vue/reactivity';
-import ComparePage from './ComparePage.vue';
-import Glossary from './Glossary/Glossary.vue';
+import { defineAsyncComponent } from 'vue';
 import HomePage from './HomePage.vue';
 import WeaponViewer from './WeaponViewer/WeaponViewer.vue';
+
+const Glossary = defineAsyncComponent(() => import("./Glossary/Glossary.vue"));
+const ComparePage = defineAsyncComponent(() => import("./ComparePage.vue"));
 
 const props = defineProps<{
     page: PageSelection,